Wednesday March 1,1905
Curiosity About the Arrival of Royal Rifles
There is a report on page three for March 1, 1905 that British troops will arrive on the coming Saturday. It also explains that they will be quartered in a nearby location. Four companies are a good amount of soldiers in a time of peace to me moving around and transporting all at once. This is very interesting because there is already troops in Alexandria so why would they need more. They are almost making sure that the city stays secure and possibly preparing for some kind of attack. The British army must feel that Alexandria is an important place to keep a close eye on and protect if anything were to happen. With more firepower in the city, the people in the city are less likely to commit crime and any outsiders would be discouraged to attack with my troops on their way over to the city. There are many possibilities for why they were moving troops near the city and they were being discrete about the movement which raises more eyebrows as well. In the article this is a smaller piece but the author may not have much information about the movement or reasons behind it.