Technical Blog Post
Technical Issues That I’ve Encountered 
During the scanning of images many problems came to my attention. The microfilm scanner was sometimes blurry and very difficult to adjust without doing various different things to it. Sometimes it would be the film that was blurry but you could not tell until you’ve tried to fix the scanner and make sure it was not the scanner. This took time to make sure you got the best images as you could. I also had lots of difficulties with the Abby fine reader as most of my images are clear, they will not register on the application. I have still been working on getting my page threes done as I am almost typing out every word in the article, while attempting to keep the correct format. That is very time consuming and I have little time left to complete it. Other technical issues include me just not knowing all the functions of GitHub and my computer in general. I do not know much about computers or the functions of computer applications so it takes time for me to understand all the tools they have to offer to make my life easier. If I had more knowledge about computers it would be much easier to figure out my technical issues.